Added Work Slice Subtasks and Details to OmniOutliner Row
## Settings
- You can now set a property for your main time tracking document in
OmniOutliner.
- You can also set a property for the path to that file.
## Work Slice Subtasks
- Instead of creating a new row with the task name, the script now
looks to see if the task name already exists as an OmniOutliner row. If
it does, the script will create a sub-task called “Work Slice”. If it
does not, the script will create it, then create the Work Slice
sub-task.
## Task Duration
- When finished, the script will calculate a duration and add it to the
appropriate column.
## More Details in the OmniOutliner Row
- Creates a link back to the OmniFocus task in notes field for the row
with the task name.
- Prompts for a note about the work slice, which is added to the notes
field of the work slice row.