Removed Modal Dialog Box
This version checks the time-tracking file for an entry that’s already
“running”. If found, depending on what you set the
endRunningTaskAutomatically property to, it will either end all running
tasks or just the ones associated with the currently selected task in
OmniFocus. The result will be that you either end a task or start a
task with this script.

Other notable changes:
- Created properties for the relevant columns in OmniOutliner tracking
document.
- Added an OmniOutliner template document to the repository.
- Algorithmic cleanup: instead of iterating through each cell every
time I wanted to do something with the columns, I refer to them
directly using “set value of cell” after verifying that it exists.
Instead of iterating through all rows checking for a name, now I use
“rows where value of cell is …” I wish I had known about this before -
AppleScript makes it easy to build collections of things.